Added experimental JEID support

+++ main/java/jaredbgreat/climaticbiome/generation/biome/LeafBiome.java
@@ -18,7 +18,7 @@ public class LeafBiome implements IBiomeSpecifier {
 	
 	
 	@Override
-	public int getBiome(ChunkTile tile) {
+	public long getBiome(ChunkTile tile) {
 		return biome;
 	}
